requirements.txt
anyio==4.12.0
attrdict==2.0.1
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.0
cssselect==1.3.0
cssutils==2.11.1
cycler==0.12.1
Cython==3.2.4
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fire==0.7.1
fonttools==4.60.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
ImageIO==2.37.2
imgaug==0.4.0
importlib_resources==6.5.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.7.5
lxml==6.0.2
matplotlib==3.9.4
more-itertools==10.8.0
networkx==3.2.1
numpy==1.26.4
opencv-contrib-python==4.6.0.66
opencv-python==4.12.0.88
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==25.0
paddleocr==2.7.0.3
paddlepaddle==3.2.2
pillow==11.3.0
mss
premailer==3.10.0
protobuf==6.33.2
pyclipper==1.3.0.post6
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.13.0
rarfile==4.2
requests==2.32.5
safetensors==0.7.0
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
six==1.17.0
soupsieve==2.8.1
termcolor==3.1.0
tifffile==2024.8.30
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
zipp==3.23.0
